+++ test/test_helper.rb
@@ -4,5 +4,6 @@ require "minitest/autorun"
 require "minitest/pride"
 
 require "webmock/minitest"
+require "mocha/mini_test"
 
 require "down"