+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Add `Down.open` for IO-like streaming, and deprecate `Down.stream` (janko-m)
+
 ## 2.2.1 (2016-06-06)
 
 * Make Down work on Windows (martinsefcik)